+++ newCommands/clone.ts
@@ -1,14 +1,14 @@
-import fs, { existsSync } from "fs";
+import { Str } from "@merrymake/utils";
+import { existsSync } from "fs";
+import { mkdir, writeFile } from "fs/promises";
 import { GIT_HOST } from "../config.js";
+import { addToExecuteQueue, finish } from "../exitMessages.js";
+import { outputGit } from "../printUtils.js";
 import { choice } from "../prompt.js";
 import { OrganizationId, PathToOrganization } from "../types.js";
 import { OrgFile, execPromise, sshReq, toSubdomain } from "../utils.js";
-import { outputGit } from "../printUtils.js";
 import { ToBeStructure, ensureGroupStructure } from "./fetch.js";
 import { listOrgs } from "./org.js";
-import { addToExecuteQueue, finish } from "../exitMessages.js";
-import { mkdir, writeFile } from "fs/promises";
-import { Str } from "@merrymake/utils";
 
 export async function do_clone(
   struct: ToBeStructure,
@@ -24,26 +24,39 @@ export async function do_clone(
       `${folderName}/.merrymake/conf.json`,
       JSON.stringify(orgFile)
     );
-    const eventsDir = `${folderName}/event-catalogue`;
+    const eventsDir = `${folderName}/event-configuration`;
     await mkdir(eventsDir, { recursive: true });
     await execPromise(`git init --initial-branch=main`, eventsDir);
     await execPromise(
       `git remote add origin "${GIT_HOST}/o${organizationId}/event-catalogue"`,
       eventsDir
     );
-    await writeFile(eventsDir + "/api.json", "{}");
-    await writeFile(eventsDir + "/cron.json", "{}");
-    const publicDir = `${folderName}/public`;
+    try {
+      await execPromise(`git pull origin main`, eventsDir);
+      await execPromise(`git branch --set-upstream-to=origin/main`, eventsDir);
+    } catch (e) {
+      if (!existsSync(eventsDir + "/api.json"))
+        await writeFile(eventsDir + "/api.json", "{}");
+      if (!existsSync(eventsDir + "/cron.json"))
+        await writeFile(eventsDir + "/cron.json", "{}");
+    }
+    const publicDir = `${folderName}/front-end`;
     await mkdir(publicDir, { recursive: true });
     await execPromise(`git init --initial-branch=main`, publicDir);
     await execPromise(
       `git remote add origin "${GIT_HOST}/o${organizationId}/public"`,
       publicDir
     );
-    await writeFile(
-      publicDir + "/index.html",
-      "<html><body>Hello, World!</body></html>"
-    );
+    try {
+      await execPromise(`git pull origin main`, publicDir);
+      await execPromise(`git branch --set-upstream-to=origin/main`, publicDir);
+    } catch (e) {
+      if (!existsSync(publicDir + "/index.html"))
+        await writeFile(
+          publicDir + "/index.html",
+          "<html><body>Hello, World!</body></html>"
+        );
+    }
     await ensureGroupStructure(
       { pathTo: new PathToOrganization(folderName), id: organizationId },
       struct
@@ -84,15 +97,17 @@ export async function checkout_org(
 
 export async function checkout() {
   try {
-    const orgs = await listOrgs();
-    return choice(
-      "Which organization would you like to clone?",
-      orgs.map((org) => ({
-        long: org.id,
-        text: `${org.name}`,
-        action: () => checkout_org(org.name, new OrganizationId(org.id)),
-      }))
-    );
+    return choice([], async () => {
+      const orgs = await listOrgs();
+      return {
+        options: orgs.map((org) => ({
+          long: org.id,
+          text: `${org.name}`,
+          action: () => checkout_org(org.name, new OrganizationId(org.id)),
+        })),
+        header: "Which organization would you like to clone?",
+      };
+    });
   } catch (e) {
     throw e;
   }
